feature(multi): leaderboard in multi + handling no connection
